requirements.txt
altair
numpy
pandas
pydeck
streamlit
PyPDF2
pinecone-client
langchain
openai
sendgrid
markdown
markdown
mistune
numpy
pandas
streamlit
unstructured-client
unstructured
openai
langchain-core 
langchain-community 
langchain-openai